Registry: Move StringLike from conversion to StrLike format_hint.
